Non-engine consumers need `InnerFlowId`
