+++ lib/opte-ioctl/src/lib.rs
@@ -16,6 +16,7 @@ use opte::api::DumpTcpFlowsReq;
 use opte::api::DumpTcpFlowsResp;
 use opte::api::DumpUftReq;
 use opte::api::DumpUftResp;
+pub use opte::api::InnerFlowId;
 use opte::api::ListLayersReq;
 use opte::api::ListLayersResp;
 use opte::api::NoResp;
@@ -24,7 +25,6 @@ use opte::api::OpteCmdIoctl;
 pub use opte::api::OpteError;
 use opte::api::SetXdeUnderlayReq;
 use opte::api::XDE_IOC_OPTE_CMD;
-use opte::engine::packet::InnerFlowId;
 use oxide_vpc::api::AddFwRuleReq;
 use oxide_vpc::api::AddRouterEntryReq;
 use oxide_vpc::api::AllowCidrReq;
