+++ lib/opte/src/api.rs
@@ -1,0 +1,149 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// Copyright 2025 Oxide Computer Company
+
+use core::fmt::Display;
+use core::fmt::{self};
+#[cfg(feature = "engine")]
+use core::hash::Hash;
+#[cfg(feature = "engine")]
+use crc32fast::Hasher;
+pub use opte_api::*;
+use serde::Deserialize;
+use serde::Serialize;
+
+const AF_INET: i32 = 2;
+const AF_INET6: i32 = 26;
+
+pub static FLOW_ID_DEFAULT: InnerFlowId = InnerFlowId {
+    proto: 255,
+    addrs: AddrPair::V4 { src: Ipv4Addr::ANY_ADDR, dst: Ipv4Addr::ANY_ADDR },
+    src_port: 0,
+    dst_port: 0,
+};
+
+/// The flow identifier.
+///
+/// In this case the flow identifier is the 5-tuple of the inner IP
+/// packet.
+///
+/// NOTE: This should not be defined in `opte`. Rather, the engine
+/// should be generic in regards to the flow identifier, and it should
+/// be up to the `NetworkImpl` to define it.
+#[derive(
+    Clone,
+    Copy,
+    Debug,
+    Deserialize,
+    Eq,
+    Hash,
+    Ord,
+    PartialEq,
+    PartialOrd,
+    Serialize,
+)]
+#[repr(C, align(4))]
+pub struct InnerFlowId {
+    // Using a `u8` here for `proto` hides the enum repr from SDTs.
+    pub proto: u8,
+    // We could also theoretically get to a 38B packing if we reduce
+    // AddrPair's repr from `u16` to `u8`. However, on the dtrace/illumos
+    // side `union addrs` is 4B aligned -- in6_addr_t has a 4B alignment.
+    // So, this layout has to match that constraint -- placing addrs at
+    // offset 0x2 with `u16` discriminant sets up 4B alignment for the
+    // enum variant data (and this struct itself is 4B aligned).
+    pub addrs: AddrPair,
+    pub src_port: u16,
+    pub dst_port: u16,
+}
+
+impl InnerFlowId {
+    #[cfg(feature = "engine")]
+    pub fn crc32(&self) -> u32 {
+        let mut hasher = Hasher::new();
+        self.hash(&mut hasher);
+        hasher.finalize()
+    }
+}
+
+impl Default for InnerFlowId {
+    fn default() -> Self {
+        FLOW_ID_DEFAULT
+    }
+}
+
+/// Tagged union of a source-dest IP address pair, used to avoid
+/// duplicating the discriminator.
+#[derive(
+    Clone,
+    Copy,
+    Debug,
+    Deserialize,
+    Eq,
+    Hash,
+    Ord,
+    PartialEq,
+    PartialOrd,
+    Serialize,
+)]
+#[repr(C, u16)]
+pub enum AddrPair {
+    V4 { src: Ipv4Addr, dst: Ipv4Addr } = AF_INET as u16,
+    V6 { src: Ipv6Addr, dst: Ipv6Addr } = AF_INET6 as u16,
+}
+
+impl AddrPair {
+    pub fn mirror(self) -> Self {
+        match self {
+            Self::V4 { src, dst } => Self::V4 { src: dst, dst: src },
+            Self::V6 { src, dst } => Self::V6 { src: dst, dst: src },
+        }
+    }
+}
+
+impl InnerFlowId {
+    /// Swap IP source and destination as well as ULP port source and
+    /// destination.
+    pub fn mirror(self) -> Self {
+        Self {
+            proto: self.proto,
+            addrs: self.addrs.mirror(),
+            src_port: self.dst_port,
+            dst_port: self.src_port,
+        }
+    }
+
+    pub fn src_ip(&self) -> IpAddr {
+        match self.addrs {
+            AddrPair::V4 { src, .. } => src.into(),
+            AddrPair::V6 { src, .. } => src.into(),
+        }
+    }
+
+    pub fn dst_ip(&self) -> IpAddr {
+        match self.addrs {
+            AddrPair::V4 { dst, .. } => dst.into(),
+            AddrPair::V6 { dst, .. } => dst.into(),
+        }
+    }
+
+    pub fn protocol(&self) -> Protocol {
+        Protocol::from(self.proto)
+    }
+}
+
+impl Display for InnerFlowId {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(
+            f,
+            "{}:{}:{}:{}:{}",
+            self.protocol(),
+            self.src_ip(),
+            self.src_port,
+            self.dst_ip(),
+            self.dst_port,
+        )
+    }
+}
